Extract general changes from update-snapshots and changes-for-nitro-2.1 branches
