+++ arbitrum-docs/node-running/how-tos/running-a-classic-node.mdx
@@ -25,7 +25,7 @@ on pre-Nitro blocks.
 ### Required artifacts
 
 - Latest Docker Image: <code>@latestClassicNodeImage@</code>
-- Latest classic snapshot for Arbitrum One: https://snapshot.arbitrum.io/mainnet/db.tar
+- Latest classic snapshot for Arbitrum One: [@arbOneClassicArchiveSnapshot@](@arbOneClassicArchiveSnapshot@)
 
 ### Required parameters
