+++ arbitrum-docs/node-running/how-tos/running-an-archive-node.mdx
@@ -35,14 +35,13 @@ The minimum storage requirements will change over time as the Nitro chain grows 
 2. **CPU:** 4+ core CPU
 3. **Storage:** Minimum 2TB SSD (make sure it is extendable)
 4. **Docker images:** We'll specify these in the below commands; you don't need to manually download them.
-   - Latest Docker image for **Arbitrum One**: <code>@latestNitroNodeImage@</code>
-   - Latest Docker image for **Arbitrum Classic**: <code>@latestClassicNodeImage@</code>
+    - Latest Docker image for **Arbitrum One**: <code>@latestNitroNodeImage@</code>
+    - Latest Docker image for **Arbitrum Classic**: <code>@latestClassicNodeImage@</code>
 5. **Database snapshots:**
-   - Arbitrum One Nitro genesis database snapshot
-     - Use the parameter `--init.url="https://snapshot.arbitrum.foundation/arb1/nitro-archive.tar"` on first startup to initialize the Nitro database
-   - Arbitrum One Classic database snapshot
-     - Download the latest Arbitrum One Classic database snapshot `"https://snapshot.arbitrum.foundation/arb1/classic-archive.tar"` to the mounted point directory.
-   - For other chains' snapshots: https://snapshot.arbitrum.foundation/index.html
+    - Nitro database snapshot
+      - Use the parameter `--init.url=` on first startup to initialize the Nitro database (you can find a list of snapshots [here](https://snapshot.arbitrum.foundation/index.html)). Example: <code>--init.url="@arbOneNitroArchiveSnapshot@"</code>
+    - Arbitrum One Classic database snapshot
+      - Download the latest Arbitrum One Classic database snapshot at [@arbOneClassicArchiveSnapshot@](@arbOneClassicArchiveSnapshot@) and place it in the mounted point directory
 
 Note that other chains don't have Classic blocks, and thus don't require an initial genesis database.
