+++ website/docusaurus.config.js
@@ -54,7 +54,7 @@ const config = {
             [
               variableInjector,
               {
-                replacements: require("./static/globalVars.js")
+                replacements: require("./src/resources/globalVars.js")
               }
             ]
           ]
