--- website/static/globalVars.js
@@ -1,7 +1,0 @@
-const globalVars = {
-    latestNitroVersionTag: 'v2.0.14',
-    latestNitroNodeImage: 'offchainlabs/nitro-node:v2.0.14-2baa834',
-    latestClassicNodeImage: 'offchainlabs/arb-node:v1.4.5-e97c1a4',
-}
-
-module.exports = globalVars;
